requirements.txt
Flask_SocketIO
Click==7.0
flask_restful
flask_cors
flask_swagger_ui
simple-websocket
Flask-Login==0.4.1
Flask-Session==0.3.1
flask-socketio
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
matplotlib
numpy>=1.20
pandas
plotly
pymongo==3.11.3
PyPortfolioOpt
python-engineio
requests
setuptools==49.6.0
six==1.11.0
statsmodels==0.11.1
Sanic
aiohttp
scipy
docopt
tabulate

